folder restruction for settings
