--- src/scenes/ProfileSettings/Item.tsx
@@ -1,10 +1,0 @@
-import { styled } from "@mui/material";
-import Paper from "@mui/material/Paper";
-
-export const Item = styled(Paper)(({ theme }: { theme: any }) => ({
-  backgroundColor: theme.palette.mode === "dark" ? "#1A2027" : "#fff",
-  ...theme.typography.body2,
-  padding: theme.spacing(1),
-  textAlign: "center",
-  color: theme.palette.text.secondary,
-}));
